:sparkles: feat : Add get list of language and grade feature #S09P12A308-236

+++ backend/airlingo/src/main/java/com/ssafy/airlingo/domain/language/repository/LanguageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.airlingo.domain.language.entity.Language;
 
-public interface LanguageRepository extends JpaRepository<Language, Integer> {
-	Language findByLanguageId(int languageId);
+public interface LanguageRepository extends JpaRepository<Language, Long> {
+	Language findByLanguageId(Long languageId);
 }
